st/nine: Explicit nine requirements
This patch raises nine requirements and disables nine for old
hw that don't match them.

It would be possible to make a lot of things work with these hw,
though not everything, but it needs special care for them in the
code, and since they are very old, it's better to drop explicitly
support for them. We are already having hard time supporting r500,
which is the most flexible dx9-only card apparently.

Signed-off-by: Axel Davy <axel.davy@ens.fr>